+++ src/components/MinimalView/MinimalSection.vue
@@ -1,29 +1,27 @@
 <template>
-   <div :class="`minimal-section-inner ${selected ? 'selected' : ''} ${showAll ? 'show-all': ''}`">
+  <div :class="`minimal-section-inner ${selected ? 'selected' : ''} ${showAll ? 'show-all': ''}`">
     <div class="section-items" v-if="items && (selected || showAll)">
-      <Item
-        v-for="(item, index) in items"
-        :id="`${index}_${makeId(item.title)}`"
-        :key="`${index}_${makeId(item.title)}`"
-        :url="item.url"
-        :title="item.title"
-        :description="item.description"
-        :icon="item.icon"
-        :target="item.target"
-        :color="item.color"
-        :backgroundColor="item.backgroundColor"
-        :statusCheckUrl="item.statusCheckUrl"
-        :statusCheckHeaders="item.statusCheckHeaders"
-        :itemSize="itemSize"
-        :hotkey="item.hotkey"
-        :enableStatusCheck="shouldEnableStatusCheck(item.statusCheck)"
-        :statusCheckAllowInsecure="item.statusCheckAllowInsecure"
-        :statusCheckAcceptCodes="item.statusCheckAcceptCodes"
-        :statusCheckMaxRedirects="item.statusCheckMaxRedirects"
-        :statusCheckInterval="getStatusCheckInterval()"
-        @itemClicked="$emit('itemClicked')"
-        @triggerModal="triggerModal"
-      />
+      <template v-for="(item) in items">
+        <SubItemGroup
+          v-if="item.subItems"
+          :key="item.id"
+          :itemId="item.id"
+          :title="item.title"
+          :subItems="item.subItems"
+          @triggerModal="triggerModal"
+        />
+        <Item
+          v-else
+          :item="item"
+          :key="item.id"
+          :itemSize="itemSize"
+          :parentSectionTitle="title"
+          @itemClicked="$emit('itemClicked')"
+          @triggerModal="triggerModal"
+          :isAddNew="false"
+          :sectionDisplayData="displayData"
+        />
+      </template>
     </div>
     <div v-if="widgets && (selected && !showAll)" class="minimal-widget-wrap">
       <WidgetBase
@@ -33,6 +31,9 @@
         :index="widgetIndx"
         @navigateToSection="navigateToSection"
       />
+    </div>
+    <div v-if="selected && !showAll && !widgets && items.length < 1" class="empty-section">
+      <p>{{ $t('home.no-items-section') }}</p>
     </div>
     <IframeModal
       :ref="`iframeModal-${groupId}`"
@@ -46,6 +47,7 @@
 import router from '@/router';
 import Item from '@/components/LinkItems/Item.vue';
 import WidgetBase from '@/components/Widgets/WidgetBase';
+import SubItemGroup from '@/components/LinkItems/SubItemGroup.vue';
 import IframeModal from '@/components/LinkItems/IframeModal.vue';
 
 export default {
@@ -71,6 +73,7 @@ export default {
   components: {
     Item,
     WidgetBase,
+    SubItemGroup,
     IframeModal,
   },
   methods: {
@@ -79,6 +82,7 @@ export default {
     },
     /* Returns a unique lowercase string, based on name, for section ID */
     makeId(str) {
+      if (!str) return 'unnamed-item';
       return str.replace(/\s+/g, '-').replace(/[^a-zA-Z ]/g, '').toLowerCase();
     },
     /* Opens the iframe modal */
@@ -101,7 +105,6 @@ export default {
       const parse = (section) => section.replace(' ', '-').toLowerCase().trim();
       const sectionIdentifier = parse(this.title);
       router.push({ path: `/home/${sectionIdentifier}` });
-      this.closeContextMenu();
     },
   },
 };
@@ -131,9 +134,18 @@ export default {
   .minimal-widget-wrap {
     padding: 1rem;
   }
+  .empty-section {
+    padding: 1rem;
+    margin: 0.5rem auto;
+    color: var(--minimal-view-group-color);
+    font-size: 1rem;
+    font-style: italic;
+    opacity: var(--dimming-factor);
+  }
   &.selected {
     border: 1px solid var(--minimal-view-group-color);
     grid-column-start: span var(--col-count, 3);
+    &:not(.show-all) { min-height: 300px; }
   }
   &.show-all {
     border: none;